Dockable windows
Todo make docking events work and move panels around

+++ project/GameProject/Core/System/Console.cpp
@@ -180,7 +180,7 @@ std::string Console::getHistory() {
 }
 
 int Console::getHistroryCount() {
-	return historyBuffer.size();
+	return (int)historyBuffer.size();
 }
 
 void Console::addToHistory(std::string h) {
